backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
pydantic==2.2.2
motor==4.3.1
python-dotenv==1.0.0
